Vašek review fixes
